+++ Game Engine/src/nullEngine/object/component/ModelComponent.java
@@ -53,7 +53,7 @@ public class ModelComponent extends GameComponent {
 	}
 
 	@Override
-	public void update(float delta, GameObject object) {
+	public void update(double delta, GameObject object) {
 
 	}
 }
