+++ Game Engine/src/nullEngine/object/component/gui/GuiComponent.java
@@ -32,7 +32,7 @@ public class GuiComponent extends GameComponent {
 	}
 
 	@Override
-	public void update(float delta, GameObject object) {
+	public void update(double delta, GameObject object) {
 
 	}
