+++ Game Engine/src/nullEngine/object/component/light/PointLight.java
@@ -25,7 +25,7 @@ public class PointLight extends GameComponent {
 	}
 
 	@Override
-	public void update(float delta, GameObject object) {
+	public void update(double delta, GameObject object) {
 
 	}
