+++ Game Engine/src/nullEngine/object/component/light/SpotLight.java
@@ -29,7 +29,7 @@ public class SpotLight extends GameComponent {
 	}
 
 	@Override
-	public void update(float delta, GameObject object) {
+	public void update(double delta, GameObject object) {
 
 	}
